Chapter7/05-ChatGPT-Movie-Recommendation/requirements.txt
line-bot-sdk==2.4.1
Flask==2.2.3
python-dotenv==0.21.1
openai==0.27.0
numpy==1.24.2
pandas==1.5.3
APScheduler==3.7.0
beautifulsoup4==4.11.1
pymongo==4.3.3